+++ Tests/FoundationEssentialsTests/HebrewCalendarTests.swift
@@ -110,98 +110,4 @@ private struct HebrewCalendarTests {
         }
     }
 
-    // MARK: - Cross-check against Foundation's ICU-backed .hebrew
-
-    // Debugging: does `Calendar(.hebrew).enumerateDates` actually work
-    // WITHOUT an explicit timezone override (i.e., system default — matches perf test).
-    @Test func debug_hanukkahEnumerateFires_systemTZ() {
-        let cal = Calendar(identifier: .hebrew)
-        print("DEBUG TZ: \(cal.timeZone.identifier)")
-        let start = Date(timeIntervalSinceReferenceDate: 496359355.795410)
-        let comps = cal.dateComponents([.year, .month, .day], from: start)
-        print("DEBUG start in \(cal.timeZone.identifier): year=\(comps.year!) month=\(comps.month!) day=\(comps.day!)")
-
-        var matches: [(Date, DateComponents)] = []
-        cal.enumerateDates(
-            startingAfter: start,
-            matching: DateComponents(month: 3, day: 25),
-            matchingPolicy: .nextTime
-        ) { result, _, stop in
-            if let d = result {
-                matches.append((d, cal.dateComponents([.year, .month, .day], from: d)))
-            }
-            if matches.count >= 3 { stop = true }
-        }
-        print("DEBUG system-TZ matches count: \(matches.count)")
-        for (d, c) in matches {
-            print("  → \(d)  heb=\(c.year!)/\(c.month!)/\(c.day!)")
-        }
-        #expect(matches.count > 0, "system-TZ enumerateDates didn't produce any matches")
-    }
-
-    // Debugging: does `Calendar(.hebrew).enumerateDates` actually work?
-    @Test func debug_hanukkahEnumerateFires() {
-        var cal = Calendar(identifier: .hebrew)
-        cal.timeZone = .gmt
-        let start = Date(timeIntervalSinceReferenceDate: 496359355.795410)
-        // Check that basic dateComponents is going through our class.
-        let comps = cal.dateComponents([.year, .month, .day], from: start)
-        print("DEBUG start \(start) → year=\(comps.year!) month=\(comps.month!) day=\(comps.day!)")
-
-        // Try enumerateDates for Hanukkah (25 Kislev).
-        var matches: [(Date, DateComponents)] = []
-        cal.enumerateDates(
-            startingAfter: start,
-            matching: DateComponents(month: 3, day: 25),
-            matchingPolicy: .nextTime
-        ) { result, exactMatch, stop in
-            if let d = result {
-                matches.append((d, cal.dateComponents([.year, .month, .day], from: d)))
-            }
-            if matches.count >= 3 { stop = true }
-        }
-        print("DEBUG matches count: \(matches.count)")
-        for (d, c) in matches {
-            print("  → \(d)  heb=\(c.year!)/\(c.month!)/\(c.day!)")
-        }
-        #expect(matches.count > 0, "enumerateDates didn't produce any Hanukkah matches")
-    }
-
-    // Compare _CalendarHebrew vs the current ICU implementation for a range of dates.
-    // Once the router is flipped, this test becomes a regression oracle against itself.
-    @Test func crossCheck_againstICU() {
-        let hebrewICU = Calendar(identifier: .hebrew)
-        // Make sure the comparison is apples-to-apples on timeZone.
-        let utcHebrewICU = {
-            var c = hebrewICU
-            c.timeZone = .gmt
-            return c
-        }()
-        let ours = makeCalendar()
-
-        // Step one day at a time for a year around ~2025.
-        var date = Date(timeIntervalSinceReferenceDate: 750000000) // 2024-09-10T04:20:00Z
-        let oneDay: TimeInterval = 86400
-
-        var divergences = 0
-        var total = 0
-        for _ in 0..<400 {
-            let icuComps = utcHebrewICU.dateComponents([.year, .month, .day], from: date)
-            let ourComps = ours.dateComponents([.year, .month, .day], from: date, in: .gmt)
-            total += 1
-            if icuComps.year != ourComps.year || icuComps.month != ourComps.month || icuComps.day != ourComps.day {
-                divergences += 1
-                if divergences <= 3 {
-                    Issue.record("Hebrew divergence at \(date): ICU=\(icuComps.year!)/\(icuComps.month!)/\(icuComps.day!) vs ours=\(ourComps.year ?? -1)/\(ourComps.month ?? -1)/\(ourComps.day ?? -1)")
-                }
-            }
-            date += oneDay
-        }
-        #expect(divergences == 0, "\(divergences)/\(total) Hebrew dates diverged from ICU")
-    }
-
-    // DST policy parity tests moved to
-    // Tests/FoundationInternationalizationTests/HebrewDSTPolicyParityTests.swift —
-    // they need `TimeZone(identifier: "America/Los_Angeles")`, which requires
-    // the ICU-backed TimeZone only linked in the Internationalization target.
 }
